+++ common/src/main/java/dev/boxadactle/coordinatesdisplay/CoordinatesDisplay.java
@@ -19,7 +19,7 @@ public class CoordinatesDisplay {
 
 	public static final String MOD_ID = "coordinatesdisplay";
 
-	public static final String VERSION = "15.0.0";
+	public static final String VERSION = "20.0.0";
 
 	public static final String VERSION_STRING = MOD_NAME + " v" + VERSION;