+++ common/src/main/java/dev/boxadactle/coordinatesdisplay/mixin/OverlayMessageTimeAccessor.java
@@ -1,11 +1,10 @@
 package dev.boxadactle.coordinatesdisplay.mixin;
 
-import net.minecraft.client.gui.Gui;
-import net.minecraft.network.chat.Component;
+import net.minecraft.client.gui.Hud;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.gen.Accessor;
 
-@Mixin(Gui.class)
+@Mixin(Hud.class)
 public interface OverlayMessageTimeAccessor {
 
     @Accessor("overlayMessageTime")