+++ CHANGELOG.md
@@ -1,3 +1,6 @@
 ## Changelog
 - Update 26.1.1
 - Fixed crash with position marking
+- Allow for position marking to be configurable (colors, text)
+- Re-organized config screen with labels
+- Removed Maximum Renderer