+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ## Changelog
-- Fixed issue where the HUD would not render
+- Fixed issue where the HUD would not render when debug was enabled