+++ CHANGELOG.md
@@ -1,2 +1,3 @@
 ## Changelog
 - The mod no longer incorrectly obtains the day counter
+- Added day time feature and time renderer